update icon: llvm (plain)
[why]
While the plain icon is plainer than the original icon, it is still
quite complex with very thin strokes.

[how]
In the discussion of PR #2264 @Snailedlt developed a more detailed plain
version of the icon.

Taking that I just removed a three-node smudge at the dragon's left hand
and did the Break-Apart & Union dance in Inkscape to clean the paths up.

Authored-by: Jørgen Kalsnes Hagen @Snailedlt
Co-authored-by: Fini Jastrow <ulf.fini.jastrow@desy.de>
Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>