update icon: nodejs (plain, plain-wordmark, line, line-wordmark) (#2418)
[why]
The plain/line item contains two lines that are not visible in the
original icon. Technically they are there but they are hidden.
These lines make it impossible to recognize the stylized N that is the
basis of the icon.

[how]
Take the rounded hexagon from the wps file (because it was the most
clean one) and hand draw the lines of the N. Create from that simple
form all the new plain/line icons.

Use a different stroke thickness for the wordmark icons because the
previously used line width is ridiculously thin.

After optimized export from Inkscape hand-edit file to re-sort elements,
remove unused fill attributes and make start identical to previous svgs.

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>
Co-authored-by: Jørgen Kalsnes Hagen <43886029+Snailedlt@users.noreply.github.com>