Sort devicons.json (as suggested by the SVG checker)
It seems not only that the current devicons.json is not sorted, it also
has un-unified indent levels and other 'defects' obviously from manual
manipulation without ever running it through a unifying process.
Which is of course not needed, but maybe it's more nice to have all
indents identical etc pp.

The commands used to create this commit:

    jq -s --indent 4 '.[] | sort_by(.name)' < devicon.json  > A
    mv A devicons.json

Signed-off-by: Fini Jastrow <ulf.fini.jastrow@desy.de>